minor warning fixes: struct vs. class forward decl; size_t vs uint
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3438 3db68df2-c116-0410-a063-a993310a9797

+++ source/renderer/VertexBuffer.h
@@ -69,7 +69,7 @@ public:
 	// return this VBs batch list
 	const std::vector<Batch*>& GetBatches() const { return m_Batches; }
 
-	const uint GetVertexSize() const { return m_VertexSize; }
+	const size_t GetVertexSize() const { return m_VertexSize; }
 
 	// free memory
 	static void Shutdown();
